@@ -14,67 +14,68 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/kadoshms/ionic2-autocomplete"
+    "url": "https://github.com/jrquick17/ionic4-auto-complete"
   },
   "author": {
-    "name": "mor",
-    "email": "kadoshms@gmail.com"
+    "name": "Jeremy Quick",
+    "email": "me@jrquick.com"
   },
   "keywords": [
     "angular"
   ],
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/kadoshms/ionic2-autocomplete/issues"
+    "url": "https://github.com/jrquick17/ionic4-auto-complete/issues"
   },
   "devDependencies": {
-    "@angular/common": "4.1.0",
-    "@angular/compiler": "4.1.0",
-    "@angular/compiler-cli": "4.1.0",
-    "@angular/core": "4.1.0",
-    "@angular/forms": "4.1.0",
-    "@angular/http": "4.1.0",
-    "@angular/platform-browser": "4.1.0",
-    "@angular/platform-browser-dynamic": "4.1.0",
-    "@compodoc/compodoc": "^1.0.0-beta.7",
-    "@ionic-native/core": "3.7.0",
-    "@ionic-native/splash-screen": "3.7.0",
-    "@ionic-native/status-bar": "3.7.0",
-    "@ionic/app-scripts": "1.3.7",
-    "@ionic/cli-plugin-cordova": "1.0.0",
-    "@ionic/cli-plugin-ionic-angular": "1.0.0",
-    "@ionic/storage": "2.0.1",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~2.0.0",
-    "core-js": "^2.4.1",
-    "del": "^2.2.2",
+    "@angular/common": "7.0.3",
+    "@angular/compiler": "7.0.3",
+    "@angular/compiler-cli": "7.0.3",
+    "@angular/core": "7.0.3",
+    "@angular/forms": "7.0.3",
+    "@angular/http": "7.0.3",
+    "@angular/platform-browser": "7.0.3",
+    "@angular/platform-browser-dynamic": "7.0.3",
+    "@compodoc/compodoc": "^1.1.6",
+    "@ionic-native/core": "5.0.0-beta.14",
+    "@ionic-native/splash-screen": "5.0.0-beta.14",
+    "@ionic-native/status-bar": "5.0.0-beta.14",
+    "@ionic/angular": "^4.0.0-beta.15",
+    "@ionic/app-scripts": "3.2.0",
+    "@ionic/cli-plugin-cordova": "1.6.2",
+    "@ionic/cli-plugin-ionic-angular": "1.4.1",
+    "@ionic/pro": "2.0.3",
+    "@ionic/storage": "2.2.0",
+    "@types/jasmine": "2.8.11",
+    "@types/node": "~10.12.7",
+    "codelyzer": "~4.5.0",
+    "core-js": "^2.5.7",
+    "del": "^3.0.0",
     "gulp": "^3.9.1",
-    "gulp-inline-source": "^3.1.0",
-    "gulp-rename": "^1.2.2",
-    "gulp-rollup": "^2.15.0",
-    "ionic-angular": "3.2.1",
-    "ionic2-auto-complete": "^1.4.3-release",
-    "ionicons": "3.0.0",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
+    "gulp-inline-source": "^3.2.0",
+    "gulp-rename": "^1.4.0",
+    "gulp-rollup": "^2.16.2",
+    "ionicons": "4.4.6",
+    "jasmine-core": "~3.3.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~3.1.1",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "node-sass": "^4.5.2",
-    "node-watch": "^0.5.2",
-    "protractor": "~5.1.0",
-    "rollup": "^0.41.6",
-    "run-sequence": "^1.2.2",
-    "rxjs": "5.1.1",
+    "karma-coverage-istanbul-reporter": "^2.0.4",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "node-sass": "^4.10.0",
+    "node-watch": "^0.5.8",
+    "protractor": "~5.4.1",
+    "rollup": "^0.67.1",
+    "run-sequence": "^2.2.1",
+    "rxjs": "6.3.3",
     "sw-toolbox": "3.6.0",
-    "ts-node": "~2.0.0",
-    "tslint": "~4.5.0",
-    "typescript": "~2.2.0",
-    "zone.js": "0.8.10"
+    "ts-node": "~7.0.1",
+    "tsickle": "^0.33.1",
+    "tslint": "~5.11.0",
+    "typescript": "~3.1.6",
+    "zone.js": "0.8.26"
   },
   "engines": {
     "node": ">=6.0.0"
