@@ -1,9 +1,9 @@
-import { NgModule, ModuleWithProviders } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule }   from '@angular/forms';
-import { AutoCompleteComponent } from './autocomplete.component';
-import { BoldPrefix } from './boldprefix.pipe';
-import {IonicModule} from 'ionic-angular';
+import {NgModule, ModuleWithProviders } from '@angular/core';
+import {CommonModule } from '@angular/common';
+import {FormsModule }   from '@angular/forms';
+import {AutoCompleteComponent } from './autocomplete.component';
+import {BoldPrefix } from './boldprefix.pipe';
+import {IonicModule} from '@ionic/angular';
 
 export * from './autocomplete.component';
 export * from './boldprefix.pipe';
